requirements.txt
sounddevice
pyttsx3
openai-whisper
torch
numpy
wave
elevenlabs
pygame

# New dependencies for enhanced features
psutil>=5.9.0
Pillow>=9.0.0
pycaw>=20220416
comtypes>=1.1.10
